move ffmpeg classes out of executor folder (executor no longer exists)
