+++ src/main/kotlin/org/jitsi/jibri/capture/ffmpeg/Commands.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  *
  */
-package org.jitsi.jibri.capture.ffmpeg.executor
+package org.jitsi.jibri.capture.ffmpeg
 
 import org.jitsi.jibri.capture.ffmpeg.FfmpegExecutorParams
 import org.jitsi.jibri.sink.Sink
