+++ src/test/kotlin/org/jitsi/jibri/capture/ffmpeg/executor/OutputParserTest.kt
@@ -21,6 +21,10 @@ import io.kotlintest.matchers.beInstanceOf
 import io.kotlintest.should
 import io.kotlintest.shouldBe
 import io.kotlintest.specs.ShouldSpec
+import org.jitsi.jibri.capture.ffmpeg.ErrorScope
+import org.jitsi.jibri.capture.ffmpeg.FfmpegErrorStatus
+import org.jitsi.jibri.capture.ffmpeg.OutputLineClassification
+import org.jitsi.jibri.capture.ffmpeg.OutputParser
 
 internal class OutputParserTest : ShouldSpec() {
     override fun isInstancePerTest(): Boolean = true
